requirements.txt
# SkillGraph Dependencies
# Core
python>=3.10

# Markdown Parsing
markdown>=3.5
PyYAML>=6.0

# Graph Processing
networkx>=3.2

# Visualization
streamlit>=1.30
pyvis>=0.3.2

# LLM Integration (optional, for future GraphRAG)
# langchain>=0.1.0
# openai>=1.0.0
# anthropic>=0.18.0

# GNN (optional, for future ML models)
# torch>=2.0
# torch-geometric>=2.4

# Development
pytest>=7.4
pytest-cov>=4.1
black>=24.0
ruff>=0.1
mypy>=1.8
